re-organize file handling, new photo_t
git-svn-id: http://svn.kfish.org/fastphoto/trunk@134 9c49b5d1-7df3-0310-bce2-b7278e68f44c

+++ src/header.c
@@ -13,9 +13,9 @@ header_content_type_jpeg ()
 }
 
 int
-header_content_length (int len)
+header_content_length (off_t len)
 {
-  return printf ("Content-Length: %d\n", len);
+  return printf ("Content-Length: %ld\n", (long)len);
 }
 
 int
